Switch to async_fetch_scoreboard_data() for all_leagues Provider
