+++ custom_components/teamtracker/coordinator.py
@@ -1,8 +1,7 @@
 """ TeamTracker Data Coordinator """
-from datetime import date, datetime, timedelta, timezone
+from datetime import datetime, timezone
 import locale
 import logging
-import re
 
 import arrow
 from async_timeout import timeout
@@ -15,7 +14,6 @@ from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, Upda
 from .clear_values import async_clear_values
 
 from .const import (
-    API_LIMIT,
     CONF_API_LANGUAGE,
     CONF_CONFERENCE_ID,
     CONF_LEAGUE_ID,
@@ -24,8 +22,6 @@ from .const import (
     CONF_TEAM_ID,
     DEFAULT_LOGO,
     DEFAULT_TIMEOUT,
-    URL_HEAD,
-    URL_TAIL,
 )
 from .event import async_process_event
 from .hockeytech import (
@@ -36,21 +32,9 @@ from .provider_factory import (
     DATA_PROVIDER_ESPN_ALL_LEAGUES, 
     DATA_PROVIDER_HOCKEYTECH, 
     get_provider)
-from .utils import is_integer, async_call_espn_api, has_team
+from .utils import is_integer
 
 _LOGGER = logging.getLogger(__name__)
-
-def _slug_to_name(slug: str) -> str:
-    """Convert a season slug like '2025-26-english-premier-league' to 'English Premier League'."""
-    if not slug:
-        return ""
-    body = re.sub(r"^\d{4}(-\d{2})?-", "", slug)
-    if body == slug:
-        return ""
-    def _fmt_word(w):
-        # Uppercase abbreviations (no vowels, e.g. "mls", "nfl"); title-case real words
-        return w.upper() if w.isalpha() and not re.search(r"[aeiou]", w, re.I) else w.title()
-    return " ".join(_fmt_word(w) for w in body.split("-"))
 
 
 class TeamTrackerCoordinator(DataUpdateCoordinator):
@@ -138,75 +122,6 @@ class TeamTrackerCoordinator(DataUpdateCoordinator):
 
         if key in TeamTrackerCoordinator.data_cache:
             TeamTrackerCoordinator.data_cache.pop(key, None)
-
-
-    #
-    #  async_get_team_schedule()
-    #
-    #    Calls the team info and schedule endpoints to discover the next game
-    #    date and build an event_id → league name mapping (substring of season)
-    #    Results are cached in all_team_cache until the next game date passes.
-    #
-    async def async_get_team_schedule(self):
-        """Fetch team schedule info for 'all' league date computation."""
-
-        team_id = self.team_id
-        sport_path = self.sport_path
-        league_path = self.league_path
-        sensor_name = self.name
-
-        cache_key = f"{sport_path}:{league_path}:{team_id}"
-        today = date.today()
-        cached = TeamTrackerCoordinator.all_team_cache.get(cache_key)
-
-        if cached is not None and today <= cached["expires"]:
-            _LOGGER.debug("%s: all_team_cache hit for '%s'", sensor_name, team_id)
-            return cached
-
-        team_url = URL_HEAD + sport_path + "/" + league_path + "/teams/" + team_id
-
-        league_map = {}
-        next_events = []
-
-        response = await async_call_espn_api(self.hass, team_url, None, sensor_name, team_id)
-        team_data = response["data"]
-        if team_data:
-            next_events = team_data.get("team", {}).get("nextEvent", [])
-            for ne in next_events:
-                eid = ne.get("id")
-                if not eid:
-                    continue
-                display = ne.get("season", {}).get("displayName") or _slug_to_name(
-                    ne.get("season", {}).get("slug", "")
-                )
-                if display:
-                    league_map[str(eid)] = display
-
-        schedule_url = team_url + "/schedule"
-        response = await async_call_espn_api(self.hass, schedule_url, None, sensor_name, team_id)
-        sched_data = response["data"]
-        if sched_data:
-            for e in sched_data.get("events", []):
-                eid = e.get("id")
-                if not eid:
-                    continue
-                display = e.get("season", {}).get("displayName") or _slug_to_name(
-                    e.get("season", {}).get("slug", "")
-                )
-                if display:
-                    league_map[str(eid)] = display
-
-        next_game_date = (
-            date.fromisoformat(next_events[0]["date"][:10]) if next_events else None
-        )
-
-        result = {
-            "next_game_date": next_game_date,
-            "league_map": league_map,
-            "expires": next_game_date or today,
-        }
-        TeamTrackerCoordinator.all_team_cache[cache_key] = result
-        return result
 
 
     #
@@ -318,80 +233,11 @@ class TeamTrackerCoordinator(DataUpdateCoordinator):
         league_path = self.league_path
         if self.provider.DATA_PROVIDER == DATA_PROVIDER_HOCKEYTECH:
             response = await async_fetch_hockeytech_data(self.hass, league_path.upper(), self.name, lang)
-        elif (league_path == "all") and is_integer(self.team_id):
-            response = await self.async_fetch_espn_all_leagues_data(self.hass, lang)
         else:
             response = await self.provider.async_fetch_scoreboard_data(self.hass, lang)
 
         self.api_url = response["url"]
         return response["data"]
-
-
-    #
-    #  async_fetch_espn_all_leagues_data()
-    #    ESPN APIs returning all leagues quickly hit the API_LIMIT, so force use of tight date ranges
-    #      1. Get the team schedule from ESPN and determine next upcoming game
-    #      2. Call w/ date range up to upcoming game
-    #      2. Call w/ date range around upcoming game
-    #
-    async def async_fetch_espn_all_leagues_data(self, hass, lang) -> dict:
-        """Gets data from ESPN APIs for all leagues in specified sport."""
-
-        sensor_name = self.name
-        sport_path = self.sport_path
-        league_path = self.league_path
-        team_id = self.team_id.upper()
-
-        # Get date of next game
-        schedule_info = await self.async_get_team_schedule()
-        next_game_date = schedule_info.get("next_game_date") if schedule_info else None
-
-        # Narrow window: cover recent results and upcoming game if within 7 days
-        today_utc = datetime.now(timezone.utc).date()
-        day_before_yesterday = today_utc - timedelta(days=2)
-
-        d1 = day_before_yesterday.strftime("%Y%m%d")
-        if next_game_date and next_game_date <= today_utc + timedelta(days=7):
-            d2 = next_game_date.strftime("%Y%m%d")
-        else:
-            d2 = today_utc.strftime("%Y%m%d")
-
-        _LOGGER.debug(
-            "%s: All-league scoreboard call 1/1 dates=%s-%s (next_game=%s)",
-            sensor_name, d1, d2,
-            next_game_date.isoformat() if next_game_date else "unknown",
-        )
-
-        url_parms = {}
-        url_parms["lang"] = lang[:2]
-        url_parms["limit"] = str(API_LIMIT)
-        url_parms["dates"] = f"{d1}-{d2}"
-
-        url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
-
-        response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id)
-        data = response["data"]
-        url = response["url"]
-
-        # If event for team not returned, narrow date range and try again
-        if has_team(data, team_id) is False:
-            if (next_game_date and next_game_date > today_utc):
-                nd1 = (next_game_date - timedelta(days=1)).strftime("%Y%m%d")
-                nd2 = next_game_date.strftime("%Y%m%d")
-                if nd1 != d1 or nd2 != d2:  # avoid duplicate call
-                    _LOGGER.debug(
-                        "%s: All-league scoreboard call 2/2 dates=%s-%s (fallback to next game)",
-                        sensor_name, nd1, nd2,
-                    )
-
-                    url_parms["dates"] = f"{nd1}-{nd2}"
-                    url = URL_HEAD + sport_path + "/" + league_path + URL_TAIL
-
-                    response = await async_call_espn_api(hass, url, url_parms, sensor_name, team_id)
-                    data = response["data"]
-                    url = response["url"]
-
-        return {"data": data, "url": url}
 
 
     #
